Move `makeCompositeAggregatePartFactory` to common
